requirements.txt
asgiref==3.5.2
Django==4.1
django-autoslug==1.9.8
django-debug-toolbar==3.5.0
django-grappelli==3.0.3
Pillow==9.2.0
pymemcache==3.5.2
pytils==0.4.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
